[exvfs] Change the interface of several exvfs_* funcs
This change makes exvfs funcs to never return directly a handle, but
to use an out-paramater instead. This allows the use of return codes
for handling several errors, typically -ENOMEM. A particularly good
example is open(): it could fail because of the file does not exist or
because there is no enough memory to allocate the file handle.

+++ kernel/char/tty.c
@@ -67,7 +67,7 @@ static const struct termios hard_coded_termios =
    0  // ospeed
 };
 
-static ssize_t tty_ioctl(fs_handle h, uptr request, void *argp)
+static int tty_ioctl(fs_handle h, uptr request, void *argp)
 {
    //printk("tty_ioctl(request: %p)\n", request);
